% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_yhat.R
\name{calc_yhat}
\alias{calc_yhat}
\title{Predictions from pgbme}
\usage{
calc_yhat(m, xInclImpList = FALSE)
}
\arguments{
\item{m}{output from pgbme}

\item{xInclImpList}{logical indicating whether imputed covariate data was used in estimation of pgbme object}
}
\description{
Takes output from pgbme and returns matrix of predicted probabilities
}
