% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplotGG.R
\name{dplotGG}
\alias{dplotGG}
\title{dplotGG}
\usage{
dplotGG(yPred, yAct, country, actual = FALSE, threshold = NULL,
  countryLabel = country)
}
\arguments{
\item{yPred}{yPred}

\item{yAct}{yAct}

\item{country}{country}

\item{actual}{actual}

\item{threshold}{threshold}

\item{countryLabel}{countryLabel}
}
\description{
dplotGG
}
