% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgPreds_pgbme.R
\name{orgPreds_pgbme}
\alias{orgPreds_pgbme}
\title{Organize predictions from pgbme}
\usage{
orgPreds_pgbme(yhat, y, directional = TRUE, threshold = NULL)
}
\arguments{
\item{yhat}{Vector of predicted probabilities}

\item{y}{Actual values}

\item{directional}{logical}

\item{threshold}{keep this set to NULL}
}
\description{
Organize predictions from pgbme
}
