% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgbme.R
\name{pgbme}
\alias{pgbme}
\title{Estimate a PGBME model}
\usage{
pgbme(y = y, Xd = array(dim = c(nrow(Y.start(y)), nrow(Y.start(y)), 0)),
  Xs = matrix(nrow = nrow(Y.start(y)), ncol = 0), Xr = matrix(nrow =
  nrow(Y.start(y)), ncol = 0), k = 1, priors = NULL, startv = NULL,
  burn = 0, pred = FALSE, rho.calc = FALSE, NS = 100, odens = 1,
  xInclImpList = FALSE, Xd_L = NULL, Xs_L = NULL, Xr_L = NULL)
}
\arguments{
\item{y}{a symmetric and unweighted square matrix (n*n).}

\item{Xd}{a n*n*rd array representing dyad-specific predictors.}

\item{Xs}{a n*rs matrix of sender-specific predictors.}

\item{Xr}{a n*rr matrix of receiver-specific predictors.}

\item{k}{an integer value indication dimension of multiplicative effect.}

\item{priors}{list of priors. Default value is NULL.}

\item{startv}{list of starting values. Default value is NULL.}

\item{burn}{size of burn-in period for MCMC.}

\item{pred}{logical indicating whether to sample from predicted distribution. Default is FALSE.}

\item{rho.calc}{logical indicating whether to estimate rho. Default is FALSE.}

\item{NS}{number of MCMC iterations.}

\item{odens}{number of iterations between saved samples.}

\item{xInclImpList}{logical indicating whether a list of imputed covariates are supplied. Default is FALSE.}

\item{Xd_L}{list of imputed dyadic covariates.}

\item{Xs_L}{list of imputed sender-specific covariates.}

\item{Xr_L}{list of imputed receiver-specific covariates.}
}
\description{
This function provides a MCMC routine to estimate a PGBME model. The model is 
detailed in "Modeling Asymmetric Relationships from Symmetric Networks" 
(Rozenas, Minhas & Ahlquist 2018).
}
