% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.y.R
\name{pred.y}
\alias{pred.y}
\title{Sample Y from posterior predictive distribution}
\usage{
pred.y(theta, rho, se, fam = "gaussian")
}
\arguments{
\item{theta}{theta}

\item{rho}{rho}

\item{se}{se}

\item{fam}{fam}
}
\description{
Sample Y from posterior predictive distribution
}
