% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbeta.d.s.gibbs.R
\name{rbeta.d.s.gibbs}
\alias{rbeta.d.s.gibbs}
\title{Sample dyadic beta and nodal effects}
\usage{
rbeta.d.s.gibbs(u, su, piS.bd, mu, s2a, rd, XTu, tXTuXTu)
}
\arguments{
\item{u}{u}

\item{su}{su}

\item{piS.bd}{piS.bd}

\item{mu}{mu}

\item{s2a}{s2a}

\item{rd}{rd}

\item{XTu}{XTu}

\item{tXTuXTu}{tXTuXTu}
}
\description{
Sample dyadic beta and nodal effects
}
