% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbeta.d.sr.gibbs.R
\name{rbeta.d.sr.gibbs}
\alias{rbeta.d.sr.gibbs}
\title{Sample sender and receiver effects}
\usage{
rbeta.d.sr.gibbs(u, v, su, sv, piS.bd, mu, Sab, n, XTu, XTv, tXTuXTu, tXTvXTv)
}
\arguments{
\item{u}{u}

\item{v}{v}

\item{su}{su}

\item{sv}{sv}

\item{piS.bd}{piS.bd}

\item{mu}{mu}

\item{Sab}{Sab}

\item{n}{n}

\item{XTu}{XTu}

\item{XTv}{XTv}

\item{tXTuXTu}{tXTuXTu}

\item{tXTvXTv}{tXTvXTv}
}
\description{
Sample sender and receiver effects
}
