% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbeta.s.gibbs.R
\name{rbeta.s.gibbs}
\alias{rbeta.s.gibbs}
\title{Pull out multivariate dist of s/r}
\usage{
rbeta.s.gibbs(s, X.u, pim.b0s, piS.b0s, s2a)
}
\arguments{
\item{s}{s}

\item{X.u}{X.u}

\item{pim.b0s}{pim.b0s}

\item{piS.b0s}{piS.b0s}

\item{s2a}{s2a}
}
\description{
Pull out multivariate dist of s/r
}
