% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbeta.sr.gibbs.R
\name{rbeta.sr.gibbs}
\alias{rbeta.sr.gibbs}
\title{Pull out multivariate dist of s/r}
\usage{
rbeta.sr.gibbs(s, r, X.u, pim.b0sr, piS.b0sr, Sab)
}
\arguments{
\item{s}{s}

\item{r}{r}

\item{X.u}{X.u}

\item{pim.b0sr}{pim.b0sr}

\item{piS.b0sr}{piS.b0sr}

\item{Sab}{Sab}
}
\description{
Pull out multivariate dist of s/r
}
