% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSab.gibbs.R
\name{rSab.gibbs}
\alias{rSab.gibbs}
\title{covariance matrix for ab}
\usage{
rSab.gibbs(a, b, S0, v0)
}
\arguments{
\item{a}{a}

\item{b}{b}

\item{S0}{S0}

\item{v0}{v0}
}
\description{
covariance matrix for ab
}
