% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rse.beta.d.gibbs.R
\name{rse.beta.d.gibbs}
\alias{rse.beta.d.gibbs}
\title{rse.beta.d.gibbs}
\usage{
rse.beta.d.gibbs(g0, g1, x, XTx, s, r, beta.d)
}
\arguments{
\item{g0}{g0}

\item{g1}{g1}

\item{x}{x}

\item{XTx}{XTx}

\item{s}{s}

\item{r}{r}

\item{beta.d}{beta.d}
}
\description{
rse.beta.d.gibbs
}
