% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta.betaX.d.srE.ef.R
\name{theta.betaX.d.srE.ef}
\alias{theta.betaX.d.srE.ef}
\title{predicted values for directed case}
\usage{
theta.betaX.d.srE.ef(beta.d, X.d, s, r, E, e, f)
}
\arguments{
\item{beta.d}{beta.d}

\item{X.d}{X.d}

\item{s}{s}

\item{r}{r}

\item{E}{E}

\item{e}{e}

\item{f}{f}
}
\description{
predicted values for directed case
}
