% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta.betaX.d.srE.z.R
\name{theta.betaX.d.srE.z}
\alias{theta.betaX.d.srE.z}
\title{predicted values for undirected case}
\usage{
theta.betaX.d.srE.z(beta.d, X.d, s, r, E, z)
}
\arguments{
\item{beta.d}{beta.d}

\item{X.d}{X.d}

\item{s}{s}

\item{r}{r}

\item{E}{E}

\item{z}{z}
}
\description{
predicted values for undirected case
}
