% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vect.mat.R
\name{vect.mat}
\alias{vect.mat}
\title{Matrix to vector}
\usage{
vect.mat(Y)
}
\arguments{
\item{Y}{a vector transformed from a matrix using mat.vect.}
}
\description{
Back-transform N*(N-1)x2 vector into NxN matrix
}
